+++ phpBB2/ctracker/admin/acp_module_changedfiles.php
@@ -38,10 +38,9 @@ if ( $action == 'akt' )
 	/*
 	 * Update the File Hashes
 	 */
+	$ct_admin->do_filechk();
 	$timestamp = time();
 	$ctracker_config->change_configuration('last_checksum_scan', $timestamp);
-	$ctracker_config->settings['last_checksum_scan'] = $timestamp;
-	$ct_admin->do_filechk();
 
 	$template->assign_block_vars('akt_complete', array(
 		'L_UPDATE_ACTION'	=> $lang['ctracker_fchk_update_action'])